requirements.txt
streamlit # the main package for making all of the site content and layout
streamlit-aggrid
streamlit_timeline

Pillow
pandas
plotly.express
json





#### You need to copy the text from the requirements-file